Add correct element clicked and navigation with args passed.

+++ app/src/main/java/com/example/napptilusandroidchallenge/oompaloompadetail/OompaLoompaDetailViewModel.kt
@@ -2,6 +2,6 @@ package com.example.napptilusandroidchallenge.oompaloompadetail
 
 import androidx.lifecycle.ViewModel
 
-class OompaLoompaDetailViewModel(oompaLoompaId: Int) : ViewModel() {
+class OompaLoompaDetailViewModel(oompaLoompaId: Long) : ViewModel() {
     // TODO: Implement the ViewModel
 }
